clean up Morrison files a bit
